requirements.txt
jinja2==2.11.3
pyyaml==5.4.1
click~=8.1.3
MarkupSafe==2.0.1

-e .